requirements.txt
pytorch                   1.9.0
torchvision               0.10.0
wandb                     0.12.1
pillow                    8.4.0 
numpy                     1.21.0
h5py                      3.3.0 
lightly                   1.1.21
gdown                     3.13.0